Add treesitter queries for outline, brackets and indents (#9)
* Add treesitter queries for outline, brackets and indents
Adds queries as defined in the zed documentation.
The outline is more or less complete for RTL code + classes, missing
covergroups and probably many other things.
I could not get `generate if/else` working properly, as the nesting in
the syntax made them display weirdly in the tree.
The `indents.scm` now makes newly-created `begin/end` blocks be
auto-indented inside.

* Fix pre-commit

---------

Co-authored-by: Owen Law <owenlaw222@gmail.com>